[#692][3.0] 피드백 반영 ############ - 불필요 파일 통합 및 코드 정리
